undrill setFilter
